Following discussion with a number of colleagues this PR covers issues #5, #6, #8, and #16.
Solutions has been used to avoid steering the group towards specific outcomes. This relates to [#5]( https://github.com/patcg/patwg-charter/issues/5).

Privacy needs to be defined and used consistently throughout and relates to [#6]( https://github.com/patcg/patwg-charter/issues/6). The PR does not address this but capitalizes the term and suggests a section to achieve this is added.

The Scope has been modified to avoid excluding solutions that address the problem the group seeks to resolve utilising a combination of different disciplines and to avoid solutions that only web browsers can implement. This relates to [#8]( https://github.com/patcg/patwg-charter/issues/8).

Regarding access to data. References the patent policy and applies access to input data under FRAND terms and principles. This will address the issue concerning data access and aligns with the existing precedent associated with those that have something that is needed to make the specification working making all those things available to others. See the new chapter on Data Policy following Patent Policy. This relates to [#16]( https://github.com/patcg/patwg-charter/issues/16).

The Success Criteria are not clear and need further work.

Included several other changes in the PR to better align the document with the W3C Process which we identified during our review. For example the call out of specific sections of the W3C Process are removed as these are not needed and emphasis some elements over others. The W3C Antitrust Guidelines are now referenced as these don’t form part of the W3C Process.

This PR is a start to move the charter in a better direction rather than a finished charter.

My colleagues advise that reviewers from Apple, Facebook, Google, and Microsoft, might wish to contact their colleagues Per Hellstrom, Tim Lamb, Oliver Bethell, and Greg Sivinski respectively for their views on the modifications.