Trade offs should be clear in the specifications
Objections included concerns that the charter did not list trade offs in utility vs privacy that would be considered during operation and consideration of specifications. Since specifications may have different concerns around utility and privacy it seems inappropriate to try and list these in the charter. Instead this change requires those concerns to be addressed on a per proposal and feature basis.