Design teams set up to address antitrust concerns
Intended to address: concerns that the charter doesn't fully comply with the W3C Antitrust and Competition Guidance. Design teams are used by the WG to lead the creation of specifications and features, adding this text is intended to assure that these teams are constructed with the antitrust concerns in mind.