Make design teams work clearly under IPR
Design teams work (like all work in the proposed WG) is under the W3C Intellectual Property Rights rules. However, this change intends to address concerns that the charter does not properly test acceptability of proposed deliverables. This change is intended to cover this concern by making it clear that we will not be producing work that isn't under PR.