Location of term definitions
This change is intended to address the concern that we might be using definitions of terms that don't match particular state, country, or industry definitions of terms, especially where those terms are defined differently by different entities. It is intended to make clear that where there are terms here that are unclear, like 'third party', based on the framework of the reader they should use the framework established by the W3C and TAG to understand those definitions. The goal is to assure all participants are on the same page.